Change partial_trace to take indicies to trace over instead of kept ones
The change is breaking, but seems like the right choice going forward.
Different libraries use different conventions, so there is no really
universal standard. The choice of supplying the indices to trace over
allows for a more straightforward implementation and removes some
ambiguity (for example, for the indices to be kept what should one do
about order? if order is not the same as original should one permute?
passing the indicies to trace over order clearly does not matter and
ordering of the dimensions is always preserved in a clear and tranparent
way).